Changes: 1)Created travis.yml file and added JaCoCo Maven plugin in pom.xml as required for the Travis/Code Climate Setup. 2)Changed radius types from int to float as specified in class on 9/19/2018. 3)Changed values in webpack.dev.config.js as suggested in Andy Dolan's piazza post to enable a "hot loading" web server. 5)Added id field in Location.java to support new field in TFFI /trip format. 4)Updated Options.java and Trip.java to support default unit options (miles, kilometers, nautical miles) and user defined miles in the TFFI format for the /trip POST request. No specifications have been given for a blank options field, therefore nothing has been added for that case.

+++ server/src/main/java/com/tripco/t13/server/Distance.java
@@ -14,7 +14,7 @@ public class Distance {
     String units;
     int distance;
 
-    public static int getDistanceNum(float theta1, float lambda1, float theta2, float lambda2, int radius) {
+    public static int getDistanceNum(float theta1, float lambda1, float theta2, float lambda2, float radius) {
 
         //convert all degree definitions to radians.
         theta1 = (float)Math.toRadians(theta1);
@@ -29,9 +29,9 @@ public class Distance {
                 Math.cos(theta1) * Math.cos(theta2) * Math.cos(lambda2 - lambda1))));
     }
 
-    public int getRadius(String units) {
+    public float getRadius(String units) {
 
-        int radius = 0;
+        float radius = 0;
         if(units.equals("miles"))
             radius = 3959;
         if(units.equals("kilometers"))
